+++ apps/storybook/stories/switch/switch-states.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Switch, _Root, _Thumb } from "@giosg/design-system-switch";
+import { Switch, _Root, _Thumb } from "@giosg-design-system/switch";
 import s from "./switch.module.css";
 
 const meta: Meta = {